# specs/006-backend-api-service/contracts/homehoney-backend-openapi.yaml
openapi: 3.1.0
info:
  title: HomeHoney Backend API
  version: 0.1.0
  summary: Unified backend API between the MAUI mobile app and FileBrowser/Mongo downstream services.
servers:
  - url: https://api.homehoney.local
    description: Local development server
paths:
...
  /api/v1/documents/insurance/{documentId}:
    get:
      summary: Get insurance document detail
      operationId: getInsuranceDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Insurance document detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsuranceRecord'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      summary: Delete insurance document
      operationId: deleteInsuranceDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiOperationResult'
  /api/v1/documents/insurance/{documentId}/files:upload:
    post:
      summary: Upload primary file for an insurance document
      operationId: uploadInsuranceDocumentFile
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Upload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiOperationResultOfInsuranceRecord'
  /api/v1/documents/insurance/{documentId}/files/{fileId}:download:
    get:
      summary: Download a file attached to an insurance document
      operationId: downloadInsuranceDocumentFile
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Binary file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /api/v1/documents/manuals/{documentId}:
    get:
      summary: Get manual document detail
      operationId: getManualDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Manual document detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualRecord'
    delete:
      summary: Delete manual document
      operationId: deleteManualDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiOperationResult'
  /api/v1/documents/manuals/{documentId}/files:upload:
    post:
      summary: Upload primary file for a manual document
      operationId: uploadManualDocumentFile
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Upload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiOperationResultOfManualRecord'
  /api/v1/documents/manuals/{documentId}/files/{fileId}:download:
    get:
      summary: Download a file attached to a manual document
      operationId: downloadManualDocumentFile
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Binary file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
components:
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
      description: FileBrowser-side file identifier. The current implementation accepts values such as `primary`.
...
  schemas:
...
    FileResource:
      type: object
      properties:
        fileResourceId:
          type: string
          format: uuid
        externalFileId:
          type: string
        externalPath:
          type: string
        fileName:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        checksumOrEtag:
          type: string
        uploadedAt:
          type: string
          format: date-time
        lastSyncedAt:
          type: string
          format: date-time
        availabilityStatus:
          type: string
          enum: [Unknown, PendingUpload, Available, Missing, SyncError]
        statusMessage:
          type: string
    InsuranceRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        policyName:
          type: string
        insuredMemberId:
          type: string
          format: uuid
        insuranceCategory:
          type: string
          enum: [Medical, CriticalIllness, Accident, Life, Other]
        status:
          type: string
          enum: [Active, PendingRenewal, Expired, Archived]
        effectiveDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        providerName:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        summary:
          type: string
        attachmentCount:
          type: integer
        primaryFile:
          $ref: '#/components/schemas/FileResource'
        integrityStatus:
          type: string
        syncMessage:
          type: string
        tags:
          type: array
          items:
            type: string
        lastUpdatedAt:
          type: string
          format: date-time
    ManualRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceName:
          type: string
        brand:
          type: string
        model:
          type: string
        spaceId:
          type: string
          format: uuid
        manualCategory:
          type: string
          enum: [Appliance, Furniture, Electronics, Kids, Other]
        purchaseDate:
          type: string
          format: date
        warrantyExpiryDate:
          type: string
          format: date
        summary:
          type: string
        attachmentCount:
          type: integer
        primaryFile:
          $ref: '#/components/schemas/FileResource'
        integrityStatus:
          type: string
        syncMessage:
          type: string
        tags:
          type: array
          items:
            type: string
        lastUpdatedAt:
          type: string
          format: date-time
...
    ApiOperationResult:
      type: object
      required: [operationId, operationType, isSuccess, code, message, occurredAt]
      properties:
        operationId:
          type: string
          format: uuid
        operationType:
          type: string
        targetType:
          type: string
        targetId:
          type: string
          format: uuid
        isSuccess:
          type: boolean
        isPartialSuccess:
          type: boolean
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        occurredAt:
          type: string
          format: date-time
...
    ApiOperationResultOfInsuranceRecord:
      allOf:
        - $ref: '#/components/schemas/ApiOperationResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/InsuranceRecord'
    ApiOperationResultOfManualRecord:
      allOf:
        - $ref: '#/components/schemas/ApiOperationResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ManualRecord'
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string